requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
channels[daphne]